Add code to implement server connections to MTS.
If you start an MTS job with MSource set to a TLNT device the
HIM emulator notices and starts a thread to listen for connections
on any of the ports MTS supports.  When a connection is received
it is assigned to a waiting MTS job.  Use *TCM in MTS to keep
some idle jobs around.